list: add new insert_and_sort_data method
When needing to maintain a sorted list,
it's often nice to be able to just sort items
as they're inserted.  This prevents a lot of
excessive looping.

This commit adds a new function that will sort
the list if it's not already sorted, and then
maintain the sort for new items added with the
function.