list: add find_node_by_comparing_data method
Right now the list finding function looks soley by
pointer comparison.  This new method allows the user
to pass in a custom comparison function.

If we know the list is sorted, we take advantage of that
by using binary search over the list to zero in on the
right node.