text-progress-bar: Try to be more defensive about text leaks
We try to prevent text from showing through in the text splash,
by disable dmesg to screen and by redirecting console messages to
a log file.

Still, occasionally a kernel message may leak through.  This
commit tries to harden against that by moving the cursor up to
home when we aren't using it.