drm: be more granular when informing kernel about flushed areas
Right now we inform the kernel about flushed areas (for the QXL)
driver a whole buffer at a time, even if we only update a small
part of the buffer.

This commit makes our flush notification much more specific, telling
the kernel exactly what part of the buffer we updated.