device-manager: drop seat abstraction in public interface
The seat abstraction turns out to be pretty wrong headed.

It forces us to make a decision on which output devices to couple
the input device to, which doesn't really make sense in the face
of multiple video cards.

This commit drops the seat object from the device manager api,
and changes the users of the api to work on the underlying
device objects directly.

Under the hood the device manager code still uses ply_seat_t.
That will get mopped up in a future commit.