device-manager: don't manage new devices if deactivated
If the drm device doesn't get loaded until the very end of the
boot process after we're deactivated, we'll still try to manage
the device and potentially step on the X servers toes.

This commit makes sure we ignore new devices when deactivated,
and potentially process them later if reactivated.