renderer: nullify device if open fails
Since commit e4f86e3c we query the renderer device straight from
the renderer plugin. Unfortunately, we continue to use that
device name even after the renderer fails to open and the plugin
gets unloaded.

This commit changes the code to go back to using the original
device name in that scenario.