wip! main: Keep detailed splash around always
If a machine has a serial console and local VT, we would ideally
show graphics on the local VT and text on the serial console.

Right now we degenerate, showing text on both.

This commit creates two splashes, one detailed and one primary,
and uses the detailed for serial consoles, and the primary
for the local vt.