ci: Fix check-format
In commit bb6580212d45f9ab0a459a9335aec2ff20818a6e we tried to stop
running check-format on debian. We did this by moving the check-format
details to its own map and then adding a "<<" merge key to only pull
it in on Fedora.

The Fedora map already has a "<<" merge key however for doing the actual
build.

It's not allowed for their to be two, and gitlab's ci just ignores
the second one.

This commit combines the merge keys into one.